repo_file and repo_dir implemented.
this implementation uses pathbuf extensively. Path cannot be used
because it is unsized type which cannot be passed around.

There is no owner for PathBuf therefore its owner is passed around. But
even if it is passed around as String instead. each function will end up
recreateing Path/PathBuf from the String.